restructure source code since place all python files in root is a bad idea
